// frontend/src/api/model/calendarResponse.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Whendoist
 * OpenAPI spec version: 0.54.72
 */

/**
 * Google Calendar with enabled status.
 */
export interface CalendarResponse {
  id: string;
  summary: string;
  primary: boolean;
  background_color: string;
  enabled: boolean;
}